Fixing integer overflow in CPoint2D

+++ Common/Header/Point2D.h
@@ -187,8 +187,8 @@ inline unsigned CPoint2D::DistanceXYZ(const CPoint2D &seg1, const CPoint2D &seg2
   int DY = seg2._y - seg1._y;
   int DZ = seg2._z - seg1._z;
   
-  double dot = X1*DX + Y1*DY + Z1*DZ;
-  double len_sq = DX*DX + DY*DY + DZ*DZ;
+  double dot = (double)X1*DX + (double)Y1*DY + (double)Z1*DZ;
+  double len_sq = (double)DX*DX + (double)DY*DY + (double)DZ*DZ;
   double param = dot / len_sq;
   
   int x, y, z;
